# docs/swagger.yaml
definitions:
  models.Artisan:
    properties:
      clusterID:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      picFormat:
        type: string
      tel:
        type: string
    type: object
  models.Cluster:
    properties:
      grade:
        type: string
      name:
        type: string
      region:
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      balance:
        type: integer
      bio:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      picFormat:
        type: string
      tel:
        type: string
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /api/v1/artisans:
    delete:
      responses: {}
      tags:
      - Artisan
    get:
      description: Show all users in the database
      responses: {}
      summary: Show all users
      tags:
      - Artisan
    post:
      parameters:
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Artisan'
      responses: {}
      tags:
      - Artisan
  /api/v1/artisans/{id}:
    delete:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - Artisan
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - Artisan
    patch:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Artisan'
      responses: {}
      tags:
      - Artisan
  /api/v1/clusters:
    delete:
      responses: {}
      tags:
      - Cluster
    get:
      description: Show all users in the database
      responses: {}
      summary: Show all users
      tags:
      - Cluster
    post:
      parameters:
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      responses: {}
      tags:
      - Cluster
  /api/v1/clusters/{id}:
    delete:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - Cluster
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - Cluster
    patch:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      responses: {}
      tags:
      - Cluster
  /api/v1/users:
    delete:
      responses: {}
      tags:
      - User
    get:
      description: Show all users in the database
      responses: {}
      summary: Show all users
      tags:
      - User
    post:
      parameters:
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses: {}
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - User
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - User
    patch:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses: {}
      tags:
      - User
swagger: "2.0"